COMMERCE-11017 Fix new test
